// dashboard/src/api/models/numberingSystemsListStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

export type NumberingSystemsListStatus =
  (typeof NumberingSystemsListStatus)[keyof typeof NumberingSystemsListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NumberingSystemsListStatus = {
  active: "active",
  archived: "archived",
} as const;